fixed float type bug

+++ movie/internal/controller/movie/controller.go
@@ -12,7 +12,7 @@ import (
 var ErrNotFound = errors.New("movie metadata not found")
 
 type ratingGateway interface {
-	GetAggregatedRating(ctx context.Context, recordID ratingmodel.RecordID, recordType ratingmodel.RecordType) (float64, error)
+	GetAggregatedRating(ctx context.Context, recordID ratingmodel.RecordID, recordType ratingmodel.RecordType) (float32, error)
 	PutRating(ctx context.Context, recordID ratingmodel.RecordID, recordType ratingmodel.RecordType, rating *ratingmodel.Rating) error
 }
